+++ src/renamer.rs
@@ -1,9 +1,12 @@
 use app::Config;
 use fileutils::{cleanup_files, create_backup, get_files};
+use std::collections::HashMap;
 use std::fs;
 use std::path::Path;
 use std::process;
 use std::sync::Arc;
+
+type RenameMap = HashMap<String, String>;
 
 pub struct Renamer {
     files: Vec<String>,
@@ -21,13 +24,44 @@ impl Renamer {
 
     /// Process file batch
     pub fn process(&mut self) {
+        let printer = &self.config.printer;
+        let colors = &printer.colors;
+
+        // Remove directories and on existing files from the list
         cleanup_files(&mut self.files);
 
-        for file in &self.files {
-            let target = self.replace_match(file);
-            if target != *file {
-                self.rename(file, &target);
+        // Relate original names with their targets
+        let rename_map = match self.get_rename_map() {
+            Ok(rename_map) => rename_map,
+            Err(_) => process::exit(1),
+        };
+
+        // Return existing targets in the list of original filenames
+        let existing_targets = match existing_targets(&rename_map) {
+            Ok(existing_targets) => existing_targets,
+            Err(err) => {
+                printer.eprint(&format!(
+                    "{}Conflict with existing file: {}",
+                    colors.error.paint("Error: "),
+                    colors.error.paint(err),
+                ));
+                process::exit(1);
             }
+        };
+
+        // Order targets to avoid renaming conflicts
+        let ordered_targets = match solve_rename_order(&rename_map, existing_targets) {
+            Ok(ordered_targets) => ordered_targets,
+            Err(err) => {
+                printer.eprint(&format!("{}{}", colors.error.paint("Error: "), err));
+                process::exit(1);
+            }
+        };
+
+        // Execute actual renaming
+        for target in &ordered_targets {
+            let source = &rename_map[target];
+            self.rename(&source, target);
         }
     }
 
@@ -46,58 +80,159 @@ impl Renamer {
         }
     }
 
-    /// Rename file in the filesystem or simply print renaming information. Checks if target
-    /// filename exists before renaming.
-    fn rename(&self, file: &str, target: &str) {
+    /// Get hash map containing all replacements to be done
+    fn get_rename_map(&self) -> Result<RenameMap, ()> {
         let printer = &self.config.printer;
         let colors = &printer.colors;
 
-        if Path::new(&target).exists() {
-            printer.print(&format!(
-                "{}File already exists - {}",
-                colors.error.paint("Error: "),
-                colors.error.paint(format!("{} -> {}", file, target))
-            ));
-        } else if self.config.force {
+        let mut rename_map = RenameMap::new();
+        let mut is_error = false;
+
+        for file in &self.files {
+            let target = self.replace_match(&file);
+            // Discard files with no changes
+            if target != *file {
+                if let Some(old_file) = rename_map.insert(target.clone(), file.clone()) {
+                    // Targets cannot be duplicated by any reason
+                    printer.eprint(&format!(
+                        "{}Two files will have the same name\n{}",
+                        colors.error.paint("Error: "),
+                        colors
+                            .error
+                            .paint(format!("{0}->{2}\n{1}->{2}", old_file, file, target))
+                    ));
+                    is_error = true;
+                }
+            }
+        }
+
+        if !is_error {
+            Ok(rename_map)
+        } else {
+            Err(())
+        }
+    }
+
+    /// Rename file in the filesystem or simply print renaming information. Checks if target
+    /// filename exists before renaming.
+    fn rename(&self, source: &str, target: &str) {
+        let printer = &self.config.printer;
+        let colors = &printer.colors;
+
+        if self.config.force {
+            // Create a backup before actual renaming
             if self.config.backup {
-                match create_backup(file) {
+                match create_backup(source) {
                     Ok(backup) => printer.print(&format!(
                         "{} Backup created - {}",
                         colors.info.paint("Info: "),
-                        colors.source.paint(format!("{} -> {}", file, backup))
+                        colors.source.paint(format!("{} -> {}", source, backup))
                     )),
                     Err(_) => {
                         printer.eprint(&format!(
                             "{}File backup failed - {}",
                             colors.error.paint("Error: "),
-                            colors.error.paint(file)
+                            colors.error.paint(source)
                         ));
                         process::exit(1);
                     }
                 }
             }
 
-            if fs::rename(&file, &target).is_err() {
+            // Rename files in the filesystem
+            if fs::rename(&source, &target).is_err() {
                 printer.eprint(&format!(
                     "{}File {} renaming failed.",
                     colors.error.paint("Error: "),
-                    colors.error.paint(file)
+                    colors.error.paint(source)
                 ));
             } else {
                 printer.print(&format!(
                     "{} -> {}",
-                    colors.source.paint(file),
+                    colors.source.paint(source),
                     colors.target.paint(target)
                 ));
             }
         } else {
+            // Just print info in dry-run mode
             printer.print(&format!(
                 "{} -> {}",
-                colors.source.paint(file),
+                colors.source.paint(source),
                 colors.target.paint(target)
             ));
         }
     }
+}
+
+/// Check if targets exists in the filesystem. If they exist, these targets must be contained in
+/// the original file list for the renaming problem to be solvable.
+fn existing_targets(rename_map: &RenameMap) -> Result<Vec<String>, String> {
+    let mut existing_targets: Vec<String> = Vec::new();
+    let sources: Vec<String> = rename_map.values().cloned().collect();
+
+    for (target, source) in rename_map {
+        if Path::new(&target).exists() {
+            if !sources.contains(&target) {
+                return Err(format!("{}->{}", source, target));
+            } else {
+                existing_targets.push(target.clone());
+            }
+        }
+    }
+    Ok(existing_targets)
+}
+
+/// Solve renaming order to avoid file overwrite. Solver will order existing targets to avoid
+/// conflicts and adds remaining targets to the list.
+fn solve_rename_order(
+    rename_map: &RenameMap,
+    mut existing_targets: Vec<String>,
+) -> Result<Vec<String>, String> {
+    // Store first all non conflicting entries
+    let mut ordered_targets: Vec<String> = rename_map
+        .iter()
+        .filter_map(|(target, _)| {
+            if !existing_targets.contains(&target) {
+                Some(target.clone())
+            } else {
+                None
+            }
+        })
+        .collect();
+
+    // Process the container with existing targets until it is empty. The algorithm extracts
+    // recursively all targets that are not present in a container with sources exclusively related
+    // to current existing targets.
+    while !existing_targets.is_empty() {
+        // Track selected index to extract value
+        let mut selected_index: Option<usize> = None;
+        // Create a vector with all sources from existing targets
+        let sources: Vec<String> = existing_targets
+            .iter()
+            .map(|x| rename_map.get(x).cloned().unwrap())
+            .collect();
+        // Select targets without conflicts in sources
+        for (index, target) in existing_targets.iter().enumerate() {
+            if sources.contains(&target) {
+                continue;
+            } else {
+                selected_index = Some(index);
+                break;
+            }
+        }
+
+        // Store result in ordered targets container or fail to stop the loop
+        match selected_index {
+            Some(index) => ordered_targets.push(existing_targets.swap_remove(index)),
+            // This will avoid infite while loop if order is not solved
+            None => return Err("Cannot solve sorting problem.".to_string()),
+        }
+    }
+
+    // Store ordered conflicting entries
+    ordered_targets.append(&mut existing_targets);
+
+    Ok(ordered_targets)
 }
 
 #[cfg(test)]